api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "django-environ>=0.12.0",
    "django-phonenumber-field[phonenumberslite]>=8.0.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "drf-spectacular>=0.28.0",
    "pillow>=11.1.0",
    "psycopg[binary]>=3.2.5",
]

[dependency-groups]
dev = [
    "ruff>=0.9.7",
]
